+++ packages/edit-post/src/store/index.js
@@ -2,12 +2,12 @@
  * WordPress dependencies
  */
 import { createReduxStore, registerStore } from '@wordpress/data';
+import { controls } from '@wordpress/data-controls';
 
 /**
  * Internal dependencies
  */
 import reducer from './reducer';
-import applyMiddlewares from './middlewares';
 import * as actions from './actions';
 import * as selectors from './selectors';
 import { STORE_NAME } from './constants';
@@ -16,6 +16,7 @@ const storeConfig = {
 	reducer,
 	actions,
 	selectors,
+	controls,
 	persist: [ 'preferences' ],
 };
 
@@ -29,6 +30,4 @@ const storeConfig = {
 export const store = createReduxStore( STORE_NAME, storeConfig );
 
 // Ideally we use register instead of register store.
-// We shouuld be able to make the switch once we remove the effects.
-const instantiatedStore = registerStore( STORE_NAME, storeConfig );
-applyMiddlewares( instantiatedStore );
+registerStore( STORE_NAME, storeConfig );